fixed pw hashing action, moved all interfaces to right folder
